rename: customer 마이페이지 네임 변경
