+++ src/app/mypagec/page.tsx
@@ -1,4 +1,4 @@
-import BannerSelfConsumer from "@/components/MyPage/customerSelf/BannerSelfConsumer";
+import BannerSelfConsumer from "@/components/MyPage/customerSelf/CustomerBannerSelf";
 import MyPageContent from "@/components/MyPage/MyPageContent";
 
 export default function MyPage(){
